orchestrator/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
redis==5.0.8
qdrant-client==1.11.1
pyyaml==6.0.2
tenacity==9.0.0
loguru==0.7.2
prometheus-client==0.20.0
python-multipart==0.0.9